Added initial audio backends.
