+++ audio-backends/null/backend.cpp
@@ -1,0 +1,29 @@
+#ifdef NULL_AUDIO
+
+#include "backend.h"
+using namespace gb;
+
+namespace gb
+{
+    void initaudio()
+    {
+
+    }
+
+    void deinitaudio()
+    {
+
+    }
+
+    void APU::mixaudio()
+    {
+
+    }
+
+    void APU::outputaudio()
+    {
+
+    }
+}
+
+#endif // NULL_AUDIO
