+++ audio-backends/null/backend.h
@@ -1,0 +1,13 @@
+#ifdef NULL_AUDIO
+
+#include "../../include/libmbGB/libmbgb.h"
+#include "../../include/libmbGB/apu.h"
+using namespace gb;
+
+namespace gb
+{
+    void initaudio();
+    void deinitaudio();
+}
+
+#endif // NULL_AUDIO
