+++ example/sdl2/main.cpp
@@ -1,5 +1,7 @@
 #include "../../include/libmbGB/libmbgb.h"
 #include <SDL2/SDL.h>
+#define SDL2_AUDIO
+#include "../../audio-backends/sdl2/backend.h"
 #include <iostream>
 #include <fstream>
 #include <sstream>
@@ -7,6 +9,7 @@
 #include <string>
 using namespace gb;
 using namespace std;
+
 
 SDL_Window *window;
 SDL_Renderer *render;
@@ -39,103 +42,9 @@ bool initSDL()
         return false;
     }
     
-    SDL_AudioSpec audiospec;
-    audiospec.format = AUDIO_F32SYS;
-    audiospec.freq = 44100;
-    audiospec.samples = 4096;
-    audiospec.channels = 2;
-    audiospec.callback = NULL;
-    
-    SDL_AudioSpec obtainedspec;
-    SDL_OpenAudio(&audiospec, &obtainedspec);
-    SDL_PauseAudio(0);
+    initsdl2audio();
     
     return true;
-}
-
-void APU::mixaudio()
-{
-    float bufferin0 = 0;
-    float bufferin1 = 0;
-    
-    int volume = (128 * leftvol) / 7;
-    
-    float sound1 = squareone.getoutputvol();
-    float sound2 = squaretwo.getoutputvol();
-    float sound3 = wave.getoutputvol();
-    float sound4 = noise.getoutputvol();
-    
-    if (leftenables[0])
-    {
-        bufferin1 = sound1;
-        SDL_MixAudioFormat((Uint8*)&bufferin0, (Uint8*)&bufferin1, AUDIO_F32SYS, sizeof(float), volume);
-    }
-    
-    if (leftenables[1])
-    {
-        bufferin1 = sound2;
-        SDL_MixAudioFormat((Uint8*)&bufferin0, (Uint8*)&bufferin1, AUDIO_F32SYS, sizeof(float), volume);
-    }
-    
-    if (leftenables[2])
-    {
-        bufferin1 = sound3;
-        SDL_MixAudioFormat((Uint8*)&bufferin0, (Uint8*)&bufferin1, AUDIO_F32SYS, sizeof(float), volume);
-    }
-    
-    if (leftenables[3])
-    {
-        bufferin1 = sound4;
-        SDL_MixAudioFormat((Uint8*)&bufferin0, (Uint8*)&bufferin1, AUDIO_F32SYS, sizeof(float), volume);
-    }
-    
-    mainbuffer[bufferfillamount] = bufferin0;
-    
-    bufferin0 = 0;
-    bufferin1 = 0;
-    
-    volume = (128 * rightvol) / 7;
-    
-    if (rightenables[0])
-    {
-        bufferin1 = sound1;
-        SDL_MixAudioFormat((Uint8*)&bufferin0, (Uint8*)&bufferin1, AUDIO_F32SYS, sizeof(float), volume);
-    }
-    
-    if (rightenables[1])
-    {
-        bufferin1 = sound2;
-        SDL_MixAudioFormat((Uint8*)&bufferin0, (Uint8*)&bufferin1, AUDIO_F32SYS, sizeof(float), volume);
-    }
-    
-    if (rightenables[2])
-    {
-        bufferin1 = sound3;
-        SDL_MixAudioFormat((Uint8*)&bufferin0, (Uint8*)&bufferin1, AUDIO_F32SYS, sizeof(float), volume);
-    }
-    
-    if (rightenables[3])
-    {
-        bufferin1 = sound4;
-        SDL_MixAudioFormat((Uint8*)&bufferin0, (Uint8*)&bufferin1, AUDIO_F32SYS, sizeof(float), volume);
-    }
-    
-    mainbuffer[bufferfillamount + 1] = bufferin0;
-    
-    bufferfillamount += 2;
-}
-
-void APU::outputaudio()
-{
-    if (bufferfillamount >= 4096)
-    {
-        bufferfillamount = 0;
-        while ((SDL_GetQueuedAudioSize(1)) > 4096 * sizeof(float))
-        {
-            SDL_Delay(1);
-        }
-        SDL_QueueAudio(1, mainbuffer, 4096 * sizeof(float));
-    }
 }
 
 void drawpixels()
@@ -162,7 +71,7 @@ void drawpixels()
 
 void stopSDL()
 {
-    SDL_CloseAudio();
+    deinitsdl2audio();
     SDL_DestroyRenderer(render);
     SDL_DestroyWindow(window);
     SDL_Quit();
@@ -171,14 +80,18 @@ void stopSDL()
 void pause()
 {
     core.paused = true;
-    SDL_PauseAudio(1);
+    #ifdef SDL2_AUDIO
+	SDL_PauseAudio(1);
+    #endif
 }
 
 void resume()
 {
     core.paused = false;
     drawpixels();
-    SDL_PauseAudio(0);
+    #ifdef SDL2_AUDIO
+	SDL_PauseAudio(0);
+    #endif
 }
 
 void changestate(int id)
@@ -294,6 +207,7 @@ int main(int argc, char* argv[])
             }
         }
         
+	#ifdef SDL2_AUDIO
         if (core.paused)
         {
             SDL_PauseAudio(1);
@@ -302,6 +216,7 @@ int main(int argc, char* argv[])
         {
             SDL_PauseAudio(0);
         }
+	#endif // SDL2_AUDIO
         
         core.runcore();
         drawpixels();
